requirements.txt
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.6
django-celery-beat==2.3.0
django-timezone-field==5.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
kombu==5.2.4
promise==2.3
prompt-toolkit==3.0.30
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
